test: performance of mkdir blocking child create
If shuffle is enabled, a parent mkdir can be placed after the creation
of its child files. While the code will still work, the process will
take *much* longer than it needs to. This tests that the placement of
this mkdir does not have an outsize effect on performance to avoid
future regressions.